test_customlist
